+++ src/yetibot/core/adapters/init.clj
@@ -1,17 +1,32 @@
 (ns yetibot.core.adapters.init
   "Manages the lifecycle of adapters"
   (:require
+    [schema.core :as s]
     [yetibot.core.adapters.slack :as slack]
     [yetibot.core.adapters.irc :as irc]
-    [taoensso.timbre :as log :refer [info warn error]]
-    [taoensso.timbre :as log :refer [info warn error]]
+    [taoensso.timbre :as log :refer [info debug warn error]]
     [clojure.stacktrace :refer [print-stack-trace]]
     [yetibot.core.adapters.adapter :as a]
-    [yetibot.core.config :refer [update-config get-config config-for-ns
-                                 reload-config conf-valid?]]
+    [yetibot.core.config :refer [get-config]]
     [yetibot.core.adapters.irc :as irc]))
 
-(defn adapters-config [] (get-config :yetibot :adapters))
+(def adapters-schema
+  {s/Keyword {:type s/Str
+              ;; different types of adapters require different keys - make them optional
+              ;; Slack
+              (s/optional-key :token) s/Str
+              ;; IRC
+              (s/optional-key :host) s/Str
+              (s/optional-key :port) s/Str
+              (s/optional-key :ssl) s/Str
+              (s/optional-key :username) s/Str
+              (s/optional-key :password) s/Str}})
+
+(defn adapters-config []
+  (let [c (get-config adapters-schema [:adapters])]
+    (if (:error c)
+      (throw (ex-info "Invalid adapters config" c))
+      (:value c))))
 
 (defn report-ex [f n]
   (try
@@ -20,11 +35,10 @@
     (catch Exception e
       (warn "Error on" n (with-out-str (print-stack-trace e))))))
 
-
-(defn make-adapter [idx config]
-  (condp = (:type config)
-    :slack (slack/make-slack idx config)
-    :irc (irc/make-irc idx config)
+(defn make-adapter [config]
+  (condp = (keyword (:type config))
+    :slack (slack/make-slack config)
+    :irc (irc/make-irc config)
     (throw (ex-info (str "Unknown adapter type " (:type config)) config))))
 
 (defn validate-adapter-config!
@@ -37,12 +51,12 @@
 
 (defn register-adapters! []
   (dorun
-    (map-indexed
-      (fn [idx adapter-config]
-        (validate-adapter-config! adapter-config)
-        (a/register-adapter!
-          (make-adapter idx adapter-config)
-          (:name adapter-config)))
+    (map
+      (fn [[uuid adapter-config]]
+        (let [adapter-config (assoc adapter-config :name uuid)]
+          (debug "Registering" (pr-str adapter-config))
+          (validate-adapter-config! adapter-config)
+          (a/register-adapter! uuid (make-adapter adapter-config))))
       (adapters-config))))
 
 (defn start-adapters! []
@@ -57,4 +71,5 @@
   (future (start-adapters!)))
 
 (defn stop []
-  (dorun (map #(a/stop %) (a/active-adapters))))
+  (dorun (map #(a/stop %) (a/active-adapters)))
+  (reset! a/adapters {}))