+++ src/yetibot/core/init.clj
@@ -4,7 +4,7 @@
     [yetibot.core.adapters.init :as ai]
     [clojure.stacktrace :refer [print-stack-trace]]
     [clojure.tools.nrepl.server :refer [start-server stop-server]]
-    [yetibot.core.config :as config]
+    [yetibot.core.config-mutable :as mconfig]
     [yetibot.core.db :as db]
     [taoensso.timbre :refer [info warn error]]
     [yetibot.core.logging :as logging]
@@ -14,14 +14,14 @@
     [yetibot.core.logo :refer [logo]])
   (:gen-class))
 
-(defn welcome-message []
+(defn welcome-message! []
   (println logo))
 
 (defn -main [& args]
   ;; only continue if able to load config
-  (if-let [c (config/reload-config!)]
+  (if-let [c (mconfig/reload-config!)]
     (do
-      (welcome-message)
+      (welcome-message!)
       (start-server :port 6789)
       (web/start-web-server)
       (db/start)
@@ -29,5 +29,5 @@
       (ai/start)
       (load-commands-and-observers))
     (do
-      (error "Yetibot failed to start: please ensure config is in place at" config/config-path " and that it is well-formed (see the log above for details)")
+      (error "Yetibot failed to start: please ensure config is in place at" mconfig/config-path " and that it is well-formed (see the log above for details)")
       (shutdown-agents))))