+++ src/yetibot/core/logging.clj
@@ -1,37 +1,40 @@
 (ns yetibot.core.logging
   (:require
+    [schema.core :as s]
     [yetibot.core.config :refer [get-config]]
     [yetibot.core.util :refer [with-fresh-db]]
     [yetibot.core.db.log :as log]
-    [taoensso.timbre.appenders.rolling :refer [make-rolling-appender]]
+    [taoensso.timbre.appenders.3rd-party.rolling :refer [rolling-appender]]
+    [taoensso.timbre.appenders.core :refer [println-appender]]
     [taoensso.timbre
      :as timbre
      :refer [trace debug info warn error fatal spy with-log-level]]))
 
-(timbre/set-level! (get-config :yetibot :logging :log-level))
+(def config-schema String)
 
-; (timbre/set-config! [:appenders :spit :enabled?] true)
-; (timbre/set-config! [:shared-appender-config :spit-filename] "/var/log/yetibot/yetibot.log")
+(defn log-level []
+  (let [v (get-config config-schema [:log :level])]
+    (if (:error v)
+      ;; default config level
+      :info
+      (keyword (:value v)))))
 
-;; rolling log files
-(timbre/set-config! [:appenders :rolling]
-                    (make-rolling-appender {:enabled? true}
-                                           {:path "/var/log/yetibot/yetibot.log"
-                                            :pattern :daily}))
-
-(defn log-to-db
-  [{:keys [ap-config level prefix throwable message] :as args}]
-  (with-fresh-db
-    (log/create (select-keys args [:level :prefix :message]))))
+(defn rolling-appender-enabled?
+  "Rolling appender is enabled by default. Disable it by setting enabled=false"
+  []
+  (let [v (get-config String [:log :rolling :enabled])]
+    (if-let [enabled (:value v)]
+      (not= enabled "false")
+      true)))
 
 (defn start []
-  ; log to datomic
-  #_(timbre/set-config!
-      [:appenders :datomic]
-      {:doc       "Datomic logger"
-       :min-level :info
-       :enabled?  true
-       :async?    false
-       :limit-per-msecs nil ; No rate limit
-       :fn #'log-to-db})
-  )
+  (timbre/set-config!
+    {:level (log-level)
+     :appenders
+     ;; stdout
+     {:println (println-appender {:stream :auto})
+      ;; rolling log files
+      :rolling-appender (rolling-appender
+                          {:enabled? (rolling-appender-enabled?)
+                           :path "/var/log/yetibot/yetibot.log"
+                           :pattern :daily})}}))